feat: Support specifying a fixed clusterIP
In some circumstances, it may be desired to give Zitadel a static
IP address instead of letting Kubernetes assign a random one. To
achieve this, the `clusterIP` field of a service with type `ClusterIP`
must be set to the desired IP address. This patch adds support for
specifying the `clusterIP` field in the `values.yaml` file of the
Zitadel chart.