chmod secrets in an emptyDir (avoiding need for root to chown to running user)
This relies on the fact that an emptyDir is set up with appropriate
permissions for whichever userid is running in the pod. The prior solution
required running chown/chmod by root in the init container. This obviates
the need for root, which is not permitted in many kubernetes installs
as a security precaution.

Fixes #177 and also fixes #178 (since there is no chown anymore).