chore: fix bump workflow for manual runs (#235)
feat: bump workflow now supports manual run with an optional ZITADEL version

Invoke this workflow manually when you want to create an updated Helm chart. If you pass in a ZITADEL release tag, then that is the version the new chart will use, otherwise the chart will automatically refer to the latest version.