feat: allow optional PAT creation for admin machine user (#420)
Adds the ability to generate a Personal Access Token (PAT) for the initial administrative machine user, in addition to the default JWT Machine Key.

This feature is opt-out. It is triggered by the presence of the FirstInstance.Org.Machine.Pat object in values.yaml. When enabled, the setup job will create a second Kubernetes secret named <username>-pat containing the PAT.

This provides users with a simpler, ready-to-use credential for basic automation and API scripting tasks immediately after installation, without removing the more secure JWT Key File option. The change is additive and fully backward-compatible.